File Logistics
Added a main_control sketch folder and file that should be the final destination for finished, working code.

Changed Critical_Subassembly to control_testing, which contains a duplicate of the main_control file for testing code. This is a redundancy, since using Git should allow for testing and version control anyway, but since we had issues last time, this should help.

File changes will be committed when this branch is merged into main.